Primera implementación con los gauges básicos

+++ Classic/ModBusLibs/modbus-tcp.c
@@ -399,8 +399,8 @@ static void _modbus_tcp_close(modbus_t *ctx)
 
 static int _modbus_tcp_flush(modbus_t *ctx)
 {
-    int rc;
-    int rc_sum = 0;
+    long rc;
+    long rc_sum = 0;
 
     do {
         /* Extract the garbage from the socket */
